+++ R/Resampling.R
@@ -18,6 +18,13 @@
 #' @template param_label
 #' @template param_man
 #'
+#' @section Stochasticity & Reproducibility:
+#' The [`Resampling`] class only defines an abstract resampling strategy.
+#' Concrete data splits are obtained by calling `$instantiate()` on a [`Task`].
+#' To ensure repdocubility of results, you need to call `set.seed` before doing so.
+#' Note that [`benchmark_grid`] internally does instantiate resamplings, so you need to set
+#' the seed before calling it.
+#'
 #' @section Stratification:
 #' All derived classes support stratified sampling.
 #' The stratification variables are assumed to be discrete and must be stored in the [Task] with column role `"stratum"`.
@@ -173,6 +180,10 @@ Resampling = R6Class("Resampling",
     #' Returns the object itself, but modified **by reference**.
     #' You need to explicitly `$clone()` the object beforehand if you want to keeps
     #' the object in its previous state.
+    #' @examples
+    #' task = tsk("penguins")
+    #' resampling = rsmp("holdout")
+    #' resampling$instantiate(task)
     instantiate = function(task) {
       task = assert_task(as_task(task))
       private$.hash = NULL
@@ -189,6 +200,10 @@ Resampling = R6Class("Resampling",
     #' @param i (`integer(1)`)\cr
     #'   Iteration.
     #' @return (`integer()`) of row ids.
+    #' @examples
+    #' task = tsk("penguins")
+    #' resampling = rsmp("holdout")$instantiate(task)
+    #' resampling$train_set(1)
     train_set = function(i) {
       private$.get_set(private$.get_train, i)
     },
@@ -199,6 +214,10 @@ Resampling = R6Class("Resampling",
     #' @param i (`integer(1)`)\cr
     #'   Iteration.
     #' @return (`integer()`) of row ids.
+    #' @examples
+    #' task = tsk("penguins")
+    #' resampling = rsmp("holdout")$instantiate(task)
+    #' resampling$test_set(1)
     test_set = function(i) {
       private$.get_set(private$.get_test, i)
     }
@@ -293,8 +312,3 @@ as.data.table.Resampling = function(x, ...) { # nolint
   set(tab, j = "set", value = factor(c("train", "test")[tab$set], levels = c("train", "test")))
   setkeyv(tab, c("set", "iteration"))[]
 }
-
-# #' @export
-# format_list_item.Resampling = function(x, ...) { # nolint
-#   sprintf("<rsmp:%s[%i]>", x$id, x$iters)
-# }